pull out introspection logic
