+++ introspection-engine/connectors/sql-introspection-connector/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod calculate_datamodel; // only exported to be able to unit test it
 mod commenting_out_guardrails;
 mod error;
+mod introspection;
 mod misc_helpers;
 mod prisma_1_defaults;
 mod re_introspection;
